requirements.txt
openai-whisper
pandas
numpy
torch
pykakasi
moviepy
wgpu-shadertoy
pillow
glfw
